test(contracts): support avs deployment fork test (#360)
Refactor AVS deployment & tests to support fork tests of same script
used for deployment.

Plus general cleanup.

task: https://app.asana.com/0/1206208509925075/1206644163432112

+++ contracts/src/interfaces/IOmniAVSAdmin.sol
@@ -82,7 +82,7 @@ interface IOmniAVSAdmin {
      * @param base The base xcall gas limit
      * @param perValidator The per-validator additional xcall gas limit
      */
-    function setXcallGasLimits(uint256 base, uint256 perValidator) external;
+    function setXcallGasLimits(uint64 base, uint64 perValidator) external;
 
     /**
      * @notice Returns true if the operator is in the allowlist.
